+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [1.19.0] - Not yet released
 
 - [bzfs] [perf] Use better heuristic for `zfs list -t snapshot` minibatch size if parallelism is disabled.
+- [bzfs] [perf] Reuse ssh connection even immediately after reboot of operating system.
 - [bzfs] Add test script that uses Lima to locally create a guest Ubuntu VM, then runs the bzfs test suite inside of
   that VM.
 - [bzfs] Also run nightly tests on AlmaLinux-10 with zfs-2.4 and `hpnssh`.